[apm] Reset outgoing HTTP connections at configurable interval
Disabled by default.

Logic shared with Logs Agent.
