+++ pkg/config/config.go
@@ -700,6 +700,7 @@ func initConfig(config Config) {
 	config.SetKnown("apm_config.service_writer.queue_size")
 	config.SetKnown("apm_config.stats_writer.connection_limit")
 	config.SetKnown("apm_config.stats_writer.queue_size")
+	config.SetKnown("apm_config.connection_reset_interval")
 	config.SetKnown("apm_config.analyzed_rate_by_service.*")
 	config.SetKnown("apm_config.analyzed_spans.*")
 	config.SetKnown("apm_config.log_throttling")
