+++ pkg/trace/writer/sender.go
@@ -23,6 +23,7 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/trace/config"
 	"github.com/DataDog/datadog-agent/pkg/trace/info"
 	"github.com/DataDog/datadog-agent/pkg/trace/osutil"
+	httputils "github.com/DataDog/datadog-agent/pkg/util/http"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
 
@@ -32,7 +33,7 @@ func newSenders(cfg *config.AgentConfig, r eventRecorder, path string, climit, q
 	if e := cfg.Endpoints; len(e) == 0 || e[0].Host == "" || e[0].APIKey == "" {
 		panic(errors.New("config was not properly validated"))
 	}
-	client := cfg.HTTPClient()
+	client := httputils.NewClient(cfg.ConnectionResetInterval, cfg.HTTPClient)
 	// spread out the the maximum connection limit (climit) between senders
 	maxConns := math.Max(1, float64(climit/len(cfg.Endpoints)))
 	senders := make([]*sender, len(cfg.Endpoints))
@@ -111,7 +112,7 @@ type eventData struct {
 // senderConfig specifies the configuration for the sender.
 type senderConfig struct {
 	// client specifies the HTTP client to use when sending requests.
-	client *http.Client
+	client *httputils.Client
 	// url specifies the URL to send requests too.
 	url *url.URL
 	// apiKey specifies the Datadog API key to use.
